+++ src/OpenLoco/include/OpenLoco/SceneManager.h
@@ -50,10 +50,8 @@ namespace OpenLoco
             progressBarActive = 1U << 4,
             initialised = 1U << 5,
             driverCheatEnabled = 1U << 6,
-            sandboxMode = 1U << 7,            // new in OpenLoco
-            pauseOverrideEnabled = 1U << 8,   // new in OpenLoco
-            ignoreIndustryObsolete = 1U << 9, // new in OpenLoco
-            generateObsoleteIndustries = 1U << 10, // new in OpenLoco
+            sandboxMode = 1U << 7,          // new in OpenLoco
+            pauseOverrideEnabled = 1U << 8, // new in OpenLoco
         };
         OPENLOCO_ENABLE_ENUM_OPERATORS(Flags);
 
@@ -91,8 +89,6 @@ namespace OpenLoco
         bool isDriverCheatEnabled();
         bool isSandboxMode();
         bool isPauseOverrideEnabled();
-        bool isIgnoreIndustryObsolete();
-        bool isGenerateObsoleteIndustries();
         bool isPaused();
         PauseFlags getPauseFlags();
         void setPauseFlag(PauseFlags value);